+++ src/js/leaflet.storage.js
@@ -74,6 +74,8 @@ L.Storage.Map.include({
         this.overrideBooleanOptionFromQueryString('datalayersControl');
         this.overrideBooleanOptionFromQueryString('displayDataBrowserOnLoad');
         this.overrideBooleanOptionFromQueryString('displayCaptionOnLoad');
+        this.datalayersOnLoad = L.Util.queryString('datalayers');
+        if (this.datalayersOnLoad) this.datalayersOnLoad = this.datalayersOnLoad.toString().split(',');
         if (L.Browser.ielt9) {
             // TODO include ie9
             this.options.allowEdit = false;
@@ -530,6 +532,7 @@ L.Storage.Map.include({
             ['dimensions.height', {handler: 'Input', label: L._('height')}],
             ['options.includeFullScreenLink', {handler: 'CheckBox', helpText: L._('Include full screen link?')}],
             ['options.currentView', {handler: 'CheckBox', helpText: L._('Current view instead of default map view?')}],
+            ['options.keepCurrentDatalayers', {handler: 'CheckBox', helpText: L._('Show current visible layers')}],
             'queryString.moreControl',
             'queryString.datalayersControl',
             'queryString.zoomControl',